fix: fix some bugs
