+++ pkg/filter/ai/kvcache/filter.go
@@ -26,9 +26,8 @@ import (
 
 	contexthttp "github.com/apache/dubbo-go-pixiu/pkg/context/http"
 	"github.com/apache/dubbo-go-pixiu/pkg/logger"
+	"github.com/go-resty/resty/v2"
 )
-
-import "github.com/go-resty/resty/v2"
 
 const (
 	Kind = constant.AIKVCacheFilter
@@ -134,37 +133,4 @@ func (f *Filter) Decode(hc *contexthttp.HttpContext) filter.FilterStatus {
 		f.manageCache(ctx, model, prompt, body, cacheStatus, routed)
 	}()
 	return filter.Continue
-}
-
-func (f *Filter) tryRouteToCachedInstance(hc *contexthttp.HttpContext, model string, prompt string) (*LookupResponse, bool) {
-	if f == nil || f.tokenManager == nil || f.lmcacheClient == nil {
-		return nil, false
-	}
-	tokens, ok := f.tokenManager.GetCachedTokens(model, prompt)
-	if !ok || len(tokens) == 0 {
-		logger.Debugf("[KVCache] routing lookup skipped: token cache miss")
-		return nil, false
-	}
-	timeout := effectiveTimeout(hc, f.cfg)
-	if f.cfg != nil && f.cfg.LookupRoutingTimeout > 0 && f.cfg.LookupRoutingTimeout < timeout {
-		timeout = f.cfg.LookupRoutingTimeout
-	}
-	ctx, cancel := context.WithTimeout(hc.Ctx, timeout)
-	defer cancel()
-	cacheStatus, err := f.lmcacheClient.Lookup(ctx, &LookupRequest{Tokens: tokens})
-	if err != nil {
-		logger.Debugf("[KVCache] routing lookup failed: %v", err)
-		return nil, false
-	}
-	instanceID := selectPreferredInstanceID(cacheStatus)
-	if instanceID == "" {
-		logger.Debugf("[KVCache] routing lookup returned empty instance")
-		return cacheStatus, false
-	}
-	if hc.Params == nil {
-		hc.Params = make(map[string]any)
-	}
-	hc.Params[constant.LLMPreferredEndpointID] = instanceID
-	logger.Debugf("[KVCache] routing preferred endpoint set: %s", instanceID)
-	return cacheStatus, true
 }
